requirements.txt
numpy==1.22.4
pandas==1.4.4
Pillow==9.2.0
tqdm==4.64.1
torch==1.12.1
torchvision==0.13.1
scikit-learn==1.1.2
matplotlib==3.5.3
seaborn==0.12.0
opencv-python==4.6.0.66
pytesseract==0.3.10
transformers==4.22.2
albumentations==1.3.0
